requirements.txt
flask==3.0.2
flask-socketio>=5.3.0
pinecone-client==3.0.2
sentence-transformers==2.5.1
python-dotenv==1.0.1
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-groq>=0.3.0
transformers==4.38.2
torch==2.2.1
torchvision==0.17.1
opencv-python==4.9.0.80
scikit-image==0.22.0 
scikit-learn>=1.3.0
pydicom>=2.4.0
reportlab>=4.0.0
eventlet>=0.35.0
gunicorn>=21.2.0
huggingface_hub>=0.20.0